+++ src/main/java/github/clone_code_detection/controller/moodle/MoodleController.java
@@ -9,8 +9,8 @@ import github.clone_code_detection.entity.moodle.dto.AssignDTO;
 import github.clone_code_detection.entity.moodle.dto.CourseDTO;
 import github.clone_code_detection.entity.moodle.dto.CourseOverviewDTO;
 import github.clone_code_detection.entity.moodle.dto.MoodleLinkRequest;
-import github.clone_code_detection.service.highlight.ServiceHighlight;
 import github.clone_code_detection.service.moodle.ServiceMoodle;
+import github.clone_code_detection.service.query.ServiceQuery;
 import org.apache.http.auth.AuthenticationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -28,29 +28,32 @@ import java.util.Collection;
 @RestController
 @RequestMapping("/api/moodle")
 public class MoodleController {
-    private final ServiceHighlight serviceHighlight;
     private final ServiceMoodle serviceMoodle;
+    private final ServiceQuery serviceQuery;
 
     @Autowired
-    public MoodleController(ServiceHighlight serviceHighlight, ServiceMoodle serviceMoodle) {
-        this.serviceHighlight = serviceHighlight;
+    public MoodleController(ServiceMoodle serviceMoodle, ServiceQuery serviceQuery) {
+        this.serviceQuery = serviceQuery;
         this.serviceMoodle = serviceMoodle;
     }
 
 
     @PostMapping(path = "/detect", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE, MediaType.APPLICATION_OCTET_STREAM_VALUE})
     @ResponseStatus(HttpStatus.OK)
-    public MoodleResponse createHighlightSessionForMoodle(@RequestParam("source") MultipartFile source)
-            throws IOException {
+    public MoodleResponse createHighlightSessionForMoodle(@RequestParam("source") MultipartFile source) throws IOException {
         // moodle source is a zip file that contains multiple folders.
         // Each folder is list of file submissions of one student.
         // For each file in student's folder, we consider it as one highlight session.
         Collection<SimilarityDetectRequest> requests = ServiceMoodle.unzipMoodleFileAndGetRequests(source);
         Collection<SimilarityReportInfoDTO> reports = new ArrayList<>();
-        requests.forEach(request -> reports.add(SimilarityReportInfoDTO.from(serviceHighlight.createSimilarityReport(
-                request,
-                IndexInstruction.getDefaultInstruction()))));
-        return MoodleResponse.builder().message("OK").data(reports).build();
+        requests.stream()
+                .map(request -> serviceQuery.createSimilarityReport(request, IndexInstruction.getDefaultInstruction()))
+                .map(SimilarityReportInfoDTO::from)
+                .forEach(reports::add);
+        return MoodleResponse.builder()
+                             .message("OK")
+                             .data(reports)
+                             .build();
     }
 
     @PostMapping(path = "/signin", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
@@ -74,9 +77,7 @@ public class MoodleController {
 
     @GetMapping(path = "/assign")
     @ResponseStatus(HttpStatus.OK)
-    public AssignDTO getAssignDetail(@RequestParam("course_id") long courseId,
-                                     @RequestParam("assign_id") long assignId,
-                                     @PageableDefault Pageable pageable) {
+    public AssignDTO getAssignDetail(@RequestParam("course_id") long courseId, @RequestParam("assign_id") long assignId, @PageableDefault Pageable pageable) {
         return serviceMoodle.getAssignDetail(courseId, assignId, pageable);
     }
